Prevent race condition
Prevent race condition of deleting/reading cache files by updateding DynamicFieldUpdate().

You can reproduce it if you installation is heavy in use. Just call AgentTicketPhone, after the original call is executed, a second ajax call is triggered. The first one is deleting the cache, the second one too which leads to an race condition, the following error message/exception is raised:

[Fri Oct 24 13:10:07 2014] [error] [client ::1] [Fri Oct 24 13:10:07 2014] index.pl: Can't cd to .. from /opt/otrs/var/tmp/CacheFileStorable/DynamicField/7/f at File/Find.pm line 983., referer: http://localhost/otrs/index.pl?Action=AgentTicketPhone

To prevent this, just update the dynamic field if needed.